+++ R/st_connect.R
@@ -57,90 +57,21 @@ st_connect = function(x, y, ids = NULL, dist, progress = TRUE, ...) {
   x = sf::st_geometry(x)
   y = sf::st_geometry(y)
 
-  # Check projection match
-  stopifnot(st_crs(x) == st_crs(y))
+  # Check that CRS is the same
+  if(!is.na(sf::st_crs(x)) & !is.na(sf::st_crs(y)) & sf::st_crs(x) != sf::st_crs(y))
+    stop("'x' and 'y' needs to be in the same CRS")
 
   # Get nearest IDs
   if(progress) cat("Calculating nearest IDs\n")
   if(is.null(ids)) ids = st_nn(x, y, progress = progress, ...)
 
-  # Numer of 'x' features
-  x_features = length(x)
-
-  # If 'x' or 'y' are polygons - convert to lines
-  if(class(x)[1] %in% c("sfc_POLYGON", "sfc_MULTIPOLYGON")) {
-    x = st_cast(x, "MULTILINESTRING")
-  }
-  if(class(y)[1] %in% c("sfc_POLYGON", "sfc_MULTIPOLYGON")) {
-    y = st_cast(y, "MULTILINESTRING")
-  }
-
-  # Calculate line lengths
-  if(class(x)[1] %in% c("sfc_LINESTRING", "sfc_MULTILINESTRING")) {
-    x_lengths = st_length(x)
-    x_lengths = as.numeric(x_lengths)
-    n_x = ceiling(x_lengths / dist)
-  }
-  if(class(y)[1] %in% c("sfc_LINESTRING", "sfc_MULTILINESTRING")) {
-    y_lengths = st_length(y)
-    y_lengths = as.numeric(y_lengths)
-    n_y = ceiling(y_lengths / dist)
-  }
-
-  # Progress bar
-  if(progress) {
-    cat("Calculating lines\n")
-    pb = utils::txtProgressBar(min = 0, max = x_features, initial = 0, style = 3)
-  }
-
   # Draw lines
-  result <- lapply(1:x_features, function(i) {
+  if(progress) cat("Calculating lines\n")
+  x = x[rep(1:length(ids), times = lengths(ids))]
+  y = y[unlist(ids)]
+  result = st_nearest_points(x, y, pairwise = TRUE)
 
-    if(class(x)[1] %in% c("sfc_LINESTRING", "sfc_MULTILINESTRING")) {
-      x_sp = as(x[i], "Spatial")
-      start_pool = sp::spsample(x_sp, type = "regular", n = n_x[i])
-      start_pool = st_as_sfc(start_pool)
-    } else {
-      start = x[i]
-    }
-
-    lines <- lapply(ids[[i]], function(j) {
-
-      if(class(x)[1] %in% c("sfc_LINESTRING", "sfc_MULTILINESTRING")) {
-        nearest_id = st_nn(y[j], start_pool, k = 1, progress = FALSE)[[1]]
-        start = start_pool[nearest_id]
-      }
-
-      if(class(y)[1] %in% c("sfc_LINESTRING", "sfc_MULTILINESTRING")) {
-        y_sp = as(y[j], "Spatial")
-        end_pool = sp::spsample(y_sp, type = "regular", n = n_y[j])
-        end_pool = st_as_sfc(end_pool)
-        nearest_id = st_nn(x[i], end_pool, k = 1, progress = FALSE)[[1]]
-        end = end_pool[nearest_id]
-      } else {
-        end = y[j]
-      }
-
-      l = c(start, end)
-      l = sf::st_combine(l)
-      l = sf::st_cast(l, "LINESTRING")
-      l
-
-    })
-
-    # Progress
-    if(progress) {
-      utils::setTxtProgressBar(pb, i)
-    }
-
-    do.call(c, lines)
-
-  })
-
-  result <- do.call(c, result)
-
-  if(progress) cat("\nDone.\n")
-
+  # Return
   return(result)
 
 }