+++ R/st_nn_pnt_proj.R
@@ -31,6 +31,10 @@
   names(ids) = NULL
   dists = split(dists, 1:nrow(dists))
   dists = lapply(dists, function(x) c(x[!is.na(x)]))
+  crs_units = st_crs(x)$units
+  dists = lapply(dists, units::set_units, crs_units, mode = "standard")
+  dists = lapply(dists, units::set_units, "m", mode = "standard")
+  dists = lapply(dists, as.numeric)
   names(dists) = NULL
 
   # Return sparse lists